[move] Node widget 用に作られた modifier を別のファイルへ移動しました.
